requirements.txt
typer
rich
pymupdf
python-docx
python-dotenv>=1.0.0
prompt_toolkit>=3.0.0
readchar>=4.0.0
pandas
ollama
neo4j
groq
